create enrollment accept, reject, checkin, cancel-checkin

+++ src/main/java/me/weekbelt/runningflex/modules/event/EventService.java
@@ -51,8 +51,26 @@ public class EventService {
     public void cancelEnrollment(Event event, Account account) {
         Enrollment enrollment = enrollmentRepository.findByEventAndAccount(event, account)
                 .orElseThrow(() -> new IllegalArgumentException("찾는 모임 신청자가 없습니다."));
-        event.removeEnrollment(enrollment);
-        enrollmentRepository.delete(enrollment);
-        event.acceptNextWaitingEnrollment();
+        if (!enrollment.isAttended()){
+            event.removeEnrollment(enrollment);
+            enrollmentRepository.delete(enrollment);
+            event.acceptNextWaitingEnrollment();
+        }
+    }
+
+    public void acceptEnrollment(Event event, Enrollment enrollment) {
+        event.accept(enrollment);
+    }
+
+    public void rejectEnrollment(Event event, Enrollment enrollment) {
+        event.reject(enrollment);
+    }
+
+    public void checkInEnrollment(Enrollment enrollment) {
+        enrollment.attended();
+    }
+
+    public void cancelCheckInEnrollment(Enrollment enrollment) {
+        enrollment.notAttended();
     }
 }
